Extract useRankString to writingStyle
