+++ src/lib/utils.ts
@@ -101,16 +101,4 @@ export function stringToBoolean(value: string): boolean {
   return value ? value.toLocaleLowerCase() === "true" : false;
 }
 
-export function useRankString(number: number): string {
-  const map = {
-    1: "main",
-    2: "second",
-    3: "third",
-    4: "fourth",
-    5: "fifth",
-  };
-
-  return map[number];
-}
-
 export type Callback<T> = (t: T) => void;
